UPD: Julia v1.6 (LTS), JuMP v0.23 (#39)
* RM: duplcate of add_ct in components.jl

* REF: move data functions to data_model

* UPD: Julia v1.6 (LTS) min

* UPD: JuMP v0.23, compat formatting

* v0.5.1 release prep

* UPD: Int64 to Int

+++ src/io/dss/dss2eng.jl
@@ -228,7 +228,7 @@ function _dss2eng_curve!(data_eng::Dict{String,<:Any}, data_dss::Dict{String,<:A
         end
         eng_obj["t_array"] = t_array
         if haskey(dss_obj, "npts")
-            npts = parse(Int64,dss_obj["npts"])
+            npts = parse(Int,dss_obj["npts"])
             if (length(c_array) != npts) || (length(t_array) != npts)
                 if length(c_array) > npts
                     @warn "c_array is longer than the npts. Truncating array."
